+++ modules/AWSPowerShell/Cmdlets/DatabaseMigrationService/Basic/Edit-DMSReplicationInstance-Cmdlet.cs
@@ -86,7 +86,11 @@ namespace Amazon.PowerShell.Cmdlets.DMS
         /// <para>A value that indicates that minor version upgrades are applied automatically to the
         /// replication instance during the maintenance window. Changing this parameter doesn't
         /// result in an outage, except in the case described following. The change is asynchronously
-        /// applied as soon as possible. </para><para>An outage does result if these factors apply: </para><ul><li><para>This parameter is set to <code>true</code> during the maintenance window.</para></li><li><para>A newer minor version is available. </para></li><li><para>DMS has enabled automatic patching for the given engine version. </para></li></ul>
+        /// applied as soon as possible. </para><para>An outage does result if these factors apply: </para><ul><li><para>This parameter is set to <code>true</code> during the maintenance window.</para></li><li><para>A newer minor version is available. </para></li><li><para>DMS has enabled automatic patching for the given engine version. </para></li></ul><para>When <code>AutoMinorVersionUpgrade</code> is enabled, DMS uses the current default
+        /// engine version when you modify a replication instance. For example, if you set <code>EngineVersion</code>
+        /// to a lower version number than the current default version, DMS uses the default version.</para><para>If <code>AutoMinorVersionUpgrade</code><i>isn’t</i> enabled when you modify a replication
+        /// instance, DMS uses the engine version specified by the <code>EngineVersion</code>
+        /// parameter.</para>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]