+++ modules/AWSPowerShell/Cmdlets/DatabaseMigrationService/Basic/Get-DMSRecommendationLimitation-Cmdlet.cs
@@ -1,0 +1,212 @@
+/*******************************************************************************
+ *  Copyright 2012-2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *  Licensed under the Apache License, Version 2.0 (the "License"). You may not use
+ *  this file except in compliance with the License. A copy of the License is located at
+ *
+ *  http://aws.amazon.com/apache2.0
+ *
+ *  or in the "license" file accompanying this file.
+ *  This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+ *  CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ *  specific language governing permissions and limitations under the License.
+ * *****************************************************************************
+ *
+ *  AWS Tools for Windows (TM) PowerShell (TM)
+ *
+ */
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Management.Automation;
+using System.Text;
+using Amazon.PowerShell.Common;
+using Amazon.Runtime;
+using Amazon.DatabaseMigrationService;
+using Amazon.DatabaseMigrationService.Model;
+
+namespace Amazon.PowerShell.Cmdlets.DMS
+{
+    /// <summary>
+    /// Returns a paginated list of limitations for recommendations of target Amazon Web Services
+    /// engines.
+    /// </summary>
+    [Cmdlet("Get", "DMSRecommendationLimitation")]
+    [OutputType("Amazon.DatabaseMigrationService.Model.Limitation")]
+    [AWSCmdlet("Calls the AWS Database Migration Service DescribeRecommendationLimitations API operation.", Operation = new[] {"DescribeRecommendationLimitations"}, SelectReturnType = typeof(Amazon.DatabaseMigrationService.Model.DescribeRecommendationLimitationsResponse))]
+    [AWSCmdletOutput("Amazon.DatabaseMigrationService.Model.Limitation or Amazon.DatabaseMigrationService.Model.DescribeRecommendationLimitationsResponse",
+        "This cmdlet returns a collection of Amazon.DatabaseMigrationService.Model.Limitation objects.",
+        "The service call response (type Amazon.DatabaseMigrationService.Model.DescribeRecommendationLimitationsResponse) can also be referenced from properties attached to the cmdlet entry in the $AWSHistory stack."
+    )]
+    public partial class GetDMSRecommendationLimitationCmdlet : AmazonDatabaseMigrationServiceClientCmdlet, IExecutor
+    {
+        
+        #region Parameter Filter
+        /// <summary>
+        /// <para>
+        /// <para>Filters applied to the limitations described in the form of key-value pairs.</para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        [Alias("Filters")]
+        public Amazon.DatabaseMigrationService.Model.Filter[] Filter { get; set; }
+        #endregion
+        
+        #region Parameter MaxRecord
+        /// <summary>
+        /// <para>
+        /// <para>The maximum number of records to include in the response. If more records exist than
+        /// the specified <code>MaxRecords</code> value, Fleet Advisor includes a pagination token
+        /// in the response so that you can retrieve the remaining results.</para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        [Alias("MaxRecords")]
+        public System.Int32? MaxRecord { get; set; }
+        #endregion
+        
+        #region Parameter NextToken
+        /// <summary>
+        /// <para>
+        /// <para>Specifies the unique pagination token that makes it possible to display the next page
+        /// of results. If this parameter is specified, the response includes only records beyond
+        /// the marker, up to the value specified by <code>MaxRecords</code>.</para><para>If <code>NextToken</code> is returned by a previous response, there are more results
+        /// available. The value of <code>NextToken</code> is a unique pagination token for each
+        /// page. Make the call again using the returned token to retrieve the next page. Keep
+        /// all other arguments unchanged.</para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        public System.String NextToken { get; set; }
+        #endregion
+        
+        #region Parameter Select
+        /// <summary>
+        /// Use the -Select parameter to control the cmdlet output. The default value is 'Limitations'.
+        /// Specifying -Select '*' will result in the cmdlet returning the whole service response (Amazon.DatabaseMigrationService.Model.DescribeRecommendationLimitationsResponse).
+        /// Specifying the name of a property of type Amazon.DatabaseMigrationService.Model.DescribeRecommendationLimitationsResponse will result in that property being returned.
+        /// Specifying -Select '^ParameterName' will result in the cmdlet returning the selected cmdlet parameter value.
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        public string Select { get; set; } = "Limitations";
+        #endregion
+        
+        protected override void ProcessRecord()
+        {
+            this._AWSSignerType = "v4";
+            base.ProcessRecord();
+            
+            var context = new CmdletContext();
+            
+            // allow for manipulation of parameters prior to loading into context
+            PreExecutionContextLoad(context);
+            
+            if (ParameterWasBound(nameof(this.Select)))
+            {
+                context.Select = CreateSelectDelegate<Amazon.DatabaseMigrationService.Model.DescribeRecommendationLimitationsResponse, GetDMSRecommendationLimitationCmdlet>(Select) ??
+                    throw new System.ArgumentException("Invalid value for -Select parameter.", nameof(this.Select));
+            }
+            if (this.Filter != null)
+            {
+                context.Filter = new List<Amazon.DatabaseMigrationService.Model.Filter>(this.Filter);
+            }
+            context.MaxRecord = this.MaxRecord;
+            context.NextToken = this.NextToken;
+            
+            // allow further manipulation of loaded context prior to processing
+            PostExecutionContextLoad(context);
+            
+            var output = Execute(context) as CmdletOutput;
+            ProcessOutput(output);
+        }
+        
+        #region IExecutor Members
+        
+        public object Execute(ExecutorContext context)
+        {
+            var cmdletContext = context as CmdletContext;
+            // create request
+            var request = new Amazon.DatabaseMigrationService.Model.DescribeRecommendationLimitationsRequest();
+            
+            if (cmdletContext.Filter != null)
+            {
+                request.Filters = cmdletContext.Filter;
+            }
+            if (cmdletContext.MaxRecord != null)
+            {
+                request.MaxRecords = cmdletContext.MaxRecord.Value;
+            }
+            if (cmdletContext.NextToken != null)
+            {
+                request.NextToken = cmdletContext.NextToken;
+            }
+            
+            CmdletOutput output;
+            
+            // issue call
+            var client = Client ?? CreateClient(_CurrentCredentials, _RegionEndpoint);
+            try
+            {
+                var response = CallAWSServiceOperation(client, request);
+                object pipelineOutput = null;
+                pipelineOutput = cmdletContext.Select(response, this);
+                output = new CmdletOutput
+                {
+                    PipelineOutput = pipelineOutput,
+                    ServiceResponse = response
+                };
+            }
+            catch (Exception e)
+            {
+                output = new CmdletOutput { ErrorResponse = e };
+            }
+            
+            return output;
+        }
+        
+        public ExecutorContext CreateContext()
+        {
+            return new CmdletContext();
+        }
+        
+        #endregion
+        
+        #region AWS Service Operation Call
+        
+        private Amazon.DatabaseMigrationService.Model.DescribeRecommendationLimitationsResponse CallAWSServiceOperation(IAmazonDatabaseMigrationService client, Amazon.DatabaseMigrationService.Model.DescribeRecommendationLimitationsRequest request)
+        {
+            Utils.Common.WriteVerboseEndpointMessage(this, client.Config, "AWS Database Migration Service", "DescribeRecommendationLimitations");
+            try
+            {
+                #if DESKTOP
+                return client.DescribeRecommendationLimitations(request);
+                #elif CORECLR
+                return client.DescribeRecommendationLimitationsAsync(request).GetAwaiter().GetResult();
+                #else
+                        #error "Unknown build edition"
+                #endif
+            }
+            catch (AmazonServiceException exc)
+            {
+                var webException = exc.InnerException as System.Net.WebException;
+                if (webException != null)
+                {
+                    throw new Exception(Utils.Common.FormatNameResolutionFailureMessage(client.Config, webException.Message), webException);
+                }
+                throw;
+            }
+        }
+        
+        #endregion
+        
+        internal partial class CmdletContext : ExecutorContext
+        {
+            public List<Amazon.DatabaseMigrationService.Model.Filter> Filter { get; set; }
+            public System.Int32? MaxRecord { get; set; }
+            public System.String NextToken { get; set; }
+            public System.Func<Amazon.DatabaseMigrationService.Model.DescribeRecommendationLimitationsResponse, GetDMSRecommendationLimitationCmdlet, object> Select { get; set; } =
+                (response, cmdlet) => response.Limitations;
+        }
+        
+    }
+}