+++ modules/AWSPowerShell/Cmdlets/DatabaseMigrationService/Basic/Set-DMSResourceTag-Cmdlet.cs
@@ -28,7 +28,7 @@ using Amazon.DatabaseMigrationService.Model;
 namespace Amazon.PowerShell.Cmdlets.DMS
 {
     /// <summary>
-    /// Adds metadata tags to an DMS resource, including replication instance, endpoint, security
+    /// Adds metadata tags to an DMS resource, including replication instance, endpoint, subnet
     /// group, and migration task. These tags can also be used with cost allocation reporting
     /// to track cost associated with DMS resources, or used in a Condition statement in an
     /// IAM policy for DMS. For more information, see <a href="https://docs.aws.amazon.com/dms/latest/APIReference/API_Tag.html"><code>Tag</code></a> data type description.