+++ modules/AWSPowerShell/Cmdlets/DatabaseMigrationService/Basic/Start-DMSBatchRecommendation-Cmdlet.cs
@@ -1,0 +1,196 @@
+/*******************************************************************************
+ *  Copyright 2012-2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *  Licensed under the Apache License, Version 2.0 (the "License"). You may not use
+ *  this file except in compliance with the License. A copy of the License is located at
+ *
+ *  http://aws.amazon.com/apache2.0
+ *
+ *  or in the "license" file accompanying this file.
+ *  This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+ *  CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ *  specific language governing permissions and limitations under the License.
+ * *****************************************************************************
+ *
+ *  AWS Tools for Windows (TM) PowerShell (TM)
+ *
+ */
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Management.Automation;
+using System.Text;
+using Amazon.PowerShell.Common;
+using Amazon.Runtime;
+using Amazon.DatabaseMigrationService;
+using Amazon.DatabaseMigrationService.Model;
+
+namespace Amazon.PowerShell.Cmdlets.DMS
+{
+    /// <summary>
+    /// Starts the analysis of up to 20 source databases to recommend target engines for each
+    /// source database. This is a batch version of <a href="https://docs.aws.amazon.com/dms/latest/APIReference/API_StartRecommendations.html">StartRecommendations</a>.
+    /// 
+    ///  
+    /// <para>
+    /// The result of analysis of each source database is reported individually in the response.
+    /// Because the batch request can result in a combination of successful and unsuccessful
+    /// actions, you should check for batch errors even when the call returns an HTTP status
+    /// code of <code>200</code>.
+    /// </para>
+    /// </summary>
+    [Cmdlet("Start", "DMSBatchRecommendation", SupportsShouldProcess = true, ConfirmImpact = ConfirmImpact.Medium)]
+    [OutputType("Amazon.DatabaseMigrationService.Model.BatchStartRecommendationsErrorEntry")]
+    [AWSCmdlet("Calls the AWS Database Migration Service BatchStartRecommendations API operation.", Operation = new[] {"BatchStartRecommendations"}, SelectReturnType = typeof(Amazon.DatabaseMigrationService.Model.BatchStartRecommendationsResponse))]
+    [AWSCmdletOutput("Amazon.DatabaseMigrationService.Model.BatchStartRecommendationsErrorEntry or Amazon.DatabaseMigrationService.Model.BatchStartRecommendationsResponse",
+        "This cmdlet returns a collection of Amazon.DatabaseMigrationService.Model.BatchStartRecommendationsErrorEntry objects.",
+        "The service call response (type Amazon.DatabaseMigrationService.Model.BatchStartRecommendationsResponse) can also be referenced from properties attached to the cmdlet entry in the $AWSHistory stack."
+    )]
+    public partial class StartDMSBatchRecommendationCmdlet : AmazonDatabaseMigrationServiceClientCmdlet, IExecutor
+    {
+        
+        #region Parameter Data
+        /// <summary>
+        /// <para>
+        /// <para>Provides information about source databases to analyze. After this analysis, Fleet
+        /// Advisor recommends target engines for each source database.</para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        public Amazon.DatabaseMigrationService.Model.StartRecommendationsRequestEntry[] Data { get; set; }
+        #endregion
+        
+        #region Parameter Select
+        /// <summary>
+        /// Use the -Select parameter to control the cmdlet output. The default value is 'ErrorEntries'.
+        /// Specifying -Select '*' will result in the cmdlet returning the whole service response (Amazon.DatabaseMigrationService.Model.BatchStartRecommendationsResponse).
+        /// Specifying the name of a property of type Amazon.DatabaseMigrationService.Model.BatchStartRecommendationsResponse will result in that property being returned.
+        /// Specifying -Select '^ParameterName' will result in the cmdlet returning the selected cmdlet parameter value.
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        public string Select { get; set; } = "ErrorEntries";
+        #endregion
+        
+        #region Parameter Force
+        /// <summary>
+        /// This parameter overrides confirmation prompts to force 
+        /// the cmdlet to continue its operation. This parameter should always
+        /// be used with caution.
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        public SwitchParameter Force { get; set; }
+        #endregion
+        
+        protected override void ProcessRecord()
+        {
+            this._AWSSignerType = "v4";
+            base.ProcessRecord();
+            
+            var resourceIdentifiersText = FormatParameterValuesForConfirmationMsg(nameof(this.Data), MyInvocation.BoundParameters);
+            if (!ConfirmShouldProceed(this.Force.IsPresent, resourceIdentifiersText, "Start-DMSBatchRecommendation (BatchStartRecommendations)"))
+            {
+                return;
+            }
+            
+            var context = new CmdletContext();
+            
+            // allow for manipulation of parameters prior to loading into context
+            PreExecutionContextLoad(context);
+            
+            if (ParameterWasBound(nameof(this.Select)))
+            {
+                context.Select = CreateSelectDelegate<Amazon.DatabaseMigrationService.Model.BatchStartRecommendationsResponse, StartDMSBatchRecommendationCmdlet>(Select) ??
+                    throw new System.ArgumentException("Invalid value for -Select parameter.", nameof(this.Select));
+            }
+            if (this.Data != null)
+            {
+                context.Data = new List<Amazon.DatabaseMigrationService.Model.StartRecommendationsRequestEntry>(this.Data);
+            }
+            
+            // allow further manipulation of loaded context prior to processing
+            PostExecutionContextLoad(context);
+            
+            var output = Execute(context) as CmdletOutput;
+            ProcessOutput(output);
+        }
+        
+        #region IExecutor Members
+        
+        public object Execute(ExecutorContext context)
+        {
+            var cmdletContext = context as CmdletContext;
+            // create request
+            var request = new Amazon.DatabaseMigrationService.Model.BatchStartRecommendationsRequest();
+            
+            if (cmdletContext.Data != null)
+            {
+                request.Data = cmdletContext.Data;
+            }
+            
+            CmdletOutput output;
+            
+            // issue call
+            var client = Client ?? CreateClient(_CurrentCredentials, _RegionEndpoint);
+            try
+            {
+                var response = CallAWSServiceOperation(client, request);
+                object pipelineOutput = null;
+                pipelineOutput = cmdletContext.Select(response, this);
+                output = new CmdletOutput
+                {
+                    PipelineOutput = pipelineOutput,
+                    ServiceResponse = response
+                };
+            }
+            catch (Exception e)
+            {
+                output = new CmdletOutput { ErrorResponse = e };
+            }
+            
+            return output;
+        }
+        
+        public ExecutorContext CreateContext()
+        {
+            return new CmdletContext();
+        }
+        
+        #endregion
+        
+        #region AWS Service Operation Call
+        
+        private Amazon.DatabaseMigrationService.Model.BatchStartRecommendationsResponse CallAWSServiceOperation(IAmazonDatabaseMigrationService client, Amazon.DatabaseMigrationService.Model.BatchStartRecommendationsRequest request)
+        {
+            Utils.Common.WriteVerboseEndpointMessage(this, client.Config, "AWS Database Migration Service", "BatchStartRecommendations");
+            try
+            {
+                #if DESKTOP
+                return client.BatchStartRecommendations(request);
+                #elif CORECLR
+                return client.BatchStartRecommendationsAsync(request).GetAwaiter().GetResult();
+                #else
+                        #error "Unknown build edition"
+                #endif
+            }
+            catch (AmazonServiceException exc)
+            {
+                var webException = exc.InnerException as System.Net.WebException;
+                if (webException != null)
+                {
+                    throw new Exception(Utils.Common.FormatNameResolutionFailureMessage(client.Config, webException.Message), webException);
+                }
+                throw;
+            }
+        }
+        
+        #endregion
+        
+        internal partial class CmdletContext : ExecutorContext
+        {
+            public List<Amazon.DatabaseMigrationService.Model.StartRecommendationsRequestEntry> Data { get; set; }
+            public System.Func<Amazon.DatabaseMigrationService.Model.BatchStartRecommendationsResponse, StartDMSBatchRecommendationCmdlet, object> Select { get; set; } =
+                (response, cmdlet) => response.ErrorEntries;
+        }
+        
+    }
+}