+++ modules/AWSPowerShell/Cmdlets/LocationService/Basic/New-LOCMap-Cmdlet.cs
@@ -114,12 +114,18 @@ namespace Amazon.PowerShell.Cmdlets.LOC
         /// the Asia Pacific (Singapore) Region (<code>ap-southeast-1</code>). For more information,
         /// see <a href="https://docs.aws.amazon.com/location/latest/developerguide/grab.html#grab-coverage-area">GrabMaps
         /// countries and area covered</a>.</para></note><para>Valid <a href="https://docs.aws.amazon.com/location/latest/developerguide/open-data.html">Open
-        /// Data (Preview) map styles</a>:</para><ul><li><para><code>VectorOpenDataStandardLight</code> – The Open Data Standard Light (preview)
-        /// map style provides a detailed basemap for the world suitable for website and mobile
-        /// application use. The map includes highways major roads, minor roads, railways, water
-        /// features, cities, parks, landmarks, building footprints, and administrative boundaries.</para><important><para>Open Data maps is in preview. We may add, change, or remove features before announcing
-        /// general availability. For more information, see <a href="https://docs.aws.amazon.com/location/latest/developerguide/open-data.html#open-data-preview">Open
-        /// Data is in preview release</a>.</para></important></li></ul>
+        /// Data map styles</a>:</para><ul><li><para><code>VectorOpenDataStandardLight</code> – The Open Data Standard Light map style
+        /// provides a detailed basemap for the world suitable for website and mobile application
+        /// use. The map includes highways major roads, minor roads, railways, water features,
+        /// cities, parks, landmarks, building footprints, and administrative boundaries.</para></li><li><para><code>VectorOpenDataStandardDark</code> – Open Data Standard Dark is a dark-themed
+        /// map style that provides a detailed basemap for the world suitable for website and
+        /// mobile application use. The map includes highways major roads, minor roads, railways,
+        /// water features, cities, parks, landmarks, building footprints, and administrative
+        /// boundaries.</para></li><li><para><code>VectorOpenDataVisualizationLight</code> – The Open Data Visualization Light
+        /// map style is a light-themed style with muted colors and fewer features that aids in
+        /// understanding overlaid data.</para></li><li><para><code>VectorOpenDataVisualizationDark</code> – The Open Data Visualization Dark map
+        /// style is a dark-themed style with muted colors and fewer features that aids in understanding
+        /// overlaid data.</para></li></ul>
         /// </para>
         /// </summary>
         #if !MODULAR