+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-﻿### 4.1.285 (2023-03-06 21:47Z)
+﻿### 4.1.286 (2023-03-07 22:03Z)
+  * AWS Tools for PowerShell now use AWS .NET SDK 3.7.498.0 and leverage its new features and improvements. Please find a description of the changes at https://github.com/aws/aws-sdk-net/blob/master/SDK.CHANGELOG.md.
+  * Amazon Database Migration Service
+    * Added cmdlet Get-DMSRecommendation leveraging the DescribeRecommendations service API.
+    * Added cmdlet Get-DMSRecommendationLimitation leveraging the DescribeRecommendationLimitations service API.
+    * Added cmdlet Start-DMSBatchRecommendation leveraging the BatchStartRecommendations service API.
+    * Added cmdlet Start-DMSRecommendation leveraging the StartRecommendations service API.
+
+### 4.1.285 (2023-03-06 21:47Z)
   * AWS Tools for PowerShell now use AWS .NET SDK 3.7.497.0 and leverage its new features and improvements. Please find a description of the changes at https://github.com/aws/aws-sdk-net/blob/master/SDK.CHANGELOG.md.
 
 ### 4.1.284 (2023-03-03 22:05Z)